fix: Phase 0 MEDIUM/LOW code fixes (0.4-0.15)
- nowUs: 64-bit multiply to avoid ~72 min overflow (main.c)
- ADC EOC: add ~10 us timeout to prevent ISR lockup (board_io.c)
- cliPrintFloat: buf[8] → buf[16] for safety (cli.c)
- EMA filter: explicit filterInitialized flag instead of value heuristic
- gain help: match code range "1..1000" (cli.c)
- CCER: read-modify-write with CCER_COMM_MASK (bldc_commutation.c)
- FaultState: remove unused struct from protection.h
- Hall_SetDirection: fix comment to match code semantics (hall.h)
- Hall_GetDriveState: mark as diagnostic-only (hall.h)
- cli.c: remove duplicate extern huart2 (already in motor_config.h)
- platformio.ini: remove meaningless build_unflags
- MOE: runtime assert after PWM init (board_io.c)

+++ src/main.c
@@ -75,7 +75,7 @@ void MotorControl_Tick(void) {
     }
 
     /* 4) Hall → komütasyon durumuna çözümleme */
-    uint32_t nowUs    = g_isrTickCount * 80U;  /* 80 µs/tick @ 12.5 kHz */
+    uint32_t nowUs    = (uint32_t)((uint64_t)g_isrTickCount * 80U);  /* 64-bit multiply to avoid ~72 min overflow */
     uint8_t  baseState = Hall_ResolveState(nowUs);
     if (baseState > 5) {
         /* Geçersiz hall — güvenli çıkış */
